Add user profile page layout
This commit creates a sample profile page using hard coded data. The data is hard coded for testing purposes, as this commit's goal is to create the layout for the profile page. Future commits will integrate this page with a backend database to store users' profiles. Data will be fetched from the backend to populate the page with information entered in by the user.