Replace localStorage for sessionStorage
This commit goes through every file and swaps localStorage for sessionStorage. The purpose of this is to allow for multiple users to be logged in on the same local device from different tabs. The decision to use localStorage was made after feedback left on a PR. SessionStorage will allow for easier testing.